codex-tools: extract utility tool specs
