+++ codex-rs/tools/src/lib.rs
@@ -5,13 +5,16 @@ mod agent_job_tool;
 mod agent_tool;
 mod code_mode;
 mod dynamic_tool;
+mod js_repl_tool;
 mod json_schema;
 mod local_tool;
+mod mcp_resource_tool;
 mod mcp_tool;
 mod request_user_input_tool;
 mod responses_api;
 mod tool_definition;
 mod tool_spec;
+mod utility_tool;
 mod view_image;
 
 pub use agent_job_tool::create_report_agent_job_result_tool;
@@ -30,8 +33,12 @@ pub use agent_tool::create_spawn_agent_tool_v2;
 pub use agent_tool::create_wait_agent_tool_v1;
 pub use agent_tool::create_wait_agent_tool_v2;
 pub use code_mode::augment_tool_spec_for_code_mode;
+pub use code_mode::create_code_mode_tool;
+pub use code_mode::create_wait_tool;
 pub use code_mode::tool_spec_to_code_mode_tool_definition;
 pub use dynamic_tool::parse_dynamic_tool;
+pub use js_repl_tool::create_js_repl_reset_tool;
+pub use js_repl_tool::create_js_repl_tool;
 pub use json_schema::AdditionalProperties;
 pub use json_schema::JsonSchema;
 pub use json_schema::parse_tool_input_schema;
@@ -42,6 +49,9 @@ pub use local_tool::create_request_permissions_tool;
 pub use local_tool::create_shell_command_tool;
 pub use local_tool::create_shell_tool;
 pub use local_tool::create_write_stdin_tool;
+pub use mcp_resource_tool::create_list_mcp_resource_templates_tool;
+pub use mcp_resource_tool::create_list_mcp_resources_tool;
+pub use mcp_resource_tool::create_read_mcp_resource_tool;
 pub use mcp_tool::mcp_call_tool_result_output_schema;
 pub use mcp_tool::parse_mcp_tool;
 pub use request_user_input_tool::create_request_user_input_tool;
@@ -61,5 +71,7 @@ pub use tool_spec::ResponsesApiWebSearchFilters;
 pub use tool_spec::ResponsesApiWebSearchUserLocation;
 pub use tool_spec::ToolSpec;
 pub use tool_spec::create_tools_json_for_responses_api;
+pub use utility_tool::create_list_dir_tool;
+pub use utility_tool::create_test_sync_tool;
 pub use view_image::ViewImageToolOptions;
 pub use view_image::create_view_image_tool;
